refator: HistoryContainer의 FlipCard 컴포넌트 분리
